MB-36085: Persist/load PINDEX_META only if manager's dataDir was set
Reviewed-on: http://review.couchbase.org/116690
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Adam Fraser <adam.fraser@couchbase.com>

+++ pindex.go
@@ -165,20 +165,23 @@ func NewPIndex(mgr *Manager, name, uuid,
 		pindex.sourcePartitionsMap[partition] = true
 	}
 
-	buf, err := json.Marshal(pindex)
-	if err != nil {
-		dest.Close()
-		os.RemoveAll(path)
-		return nil, err
-	}
+	// persist PINDEX_META only if manager's dataDir is set
+	if mgr != nil && len(mgr.dataDir) > 0 {
+		buf, err := json.Marshal(pindex)
+		if err != nil {
+			dest.Close()
+			os.RemoveAll(path)
+			return nil, err
+		}
 
-	err = ioutil.WriteFile(path+string(os.PathSeparator)+PINDEX_META_FILENAME,
-		buf, 0600)
-	if err != nil {
-		dest.Close()
-		os.RemoveAll(path)
-		return nil, fmt.Errorf("pindex: could not save PINDEX_META_FILENAME,"+
-			" path: %s, err: %v", path, err)
+		err = ioutil.WriteFile(path+string(os.PathSeparator)+PINDEX_META_FILENAME,
+			buf, 0600)
+		if err != nil {
+			dest.Close()
+			os.RemoveAll(path)
+			return nil, fmt.Errorf("pindex: could not save PINDEX_META_FILENAME,"+
+				" path: %s, err: %v", path, err)
+		}
 	}
 
 	return pindex, nil
@@ -187,18 +190,21 @@ func NewPIndex(mgr *Manager, name, uuid,
 // OpenPIndex reopens a previously created pindex.  The path argument
 // must be a directory for the pindex.
 func OpenPIndex(mgr *Manager, path string) (pindex *PIndex, err error) {
-	buf, err := ioutil.ReadFile(path +
-		string(os.PathSeparator) + PINDEX_META_FILENAME)
-	if err != nil {
-		return nil, fmt.Errorf("pindex: could not load PINDEX_META_FILENAME,"+
-			" path: %s, err: %v", path, err)
-	}
-
 	pindex = &PIndex{}
-	err = json.Unmarshal(buf, pindex)
-	if err != nil {
-		return nil, fmt.Errorf("pindex: could not parse pindex json,"+
-			" path: %s, err: %v", path, err)
+	// load PINDEX_META only if manager's dataDir is set
+	if mgr != nil && len(mgr.dataDir) > 0 {
+		buf, err := ioutil.ReadFile(path +
+			string(os.PathSeparator) + PINDEX_META_FILENAME)
+		if err != nil {
+			return nil, fmt.Errorf("pindex: could not load PINDEX_META_FILENAME,"+
+				" path: %s, err: %v", path, err)
+		}
+
+		err = json.Unmarshal(buf, pindex)
+		if err != nil {
+			return nil, fmt.Errorf("pindex: could not parse pindex json,"+
+				" path: %s, err: %v", path, err)
+		}
 	}
 
 	restart := func() {
